update example to work on mobile
