+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["wire", "devtools", "tauri-plugin", "examples/*"]
+members = ["wire", "devtools", "tauri-plugin", "examples/*/src-tauri"]
 
 [workspace.package]
 authors = ["CrabNebula <hello@crabnebula.dev>"]
