--- examples/tauri/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "example-tauri"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[[bin]]
-name = "tauri-example"
-path = "main.rs"
-
-[dependencies]
-tauri.workspace = true
-tauri-plugin-devtools = { path = "../../tauri-plugin" }
-tracing.workspace = true
-tokio = { workspace = true, features = ["time"] }
-reqwest = "0.11"
-tiny_http = "0.11"
-serde = "1"
-
-[build-dependencies]
-tauri-build.workspace = true
-
-[features]
-custom-protocol = [ "tauri/custom-protocol" ]
