+++ examples/tauri/README.md
@@ -1,3 +1,17 @@
 # Hello World Example
 
-To execute run the following: `cargo run -- --inspect`.
+To execute run the following: `cargo run`.
+
+## Android
+
+```sh
+pnpm tauri android init
+pnpm tauri android dev
+```
+
+## iOS
+
+```sh
+pnpm tauri ios init
+pnpm tauri ios dev
+```
