--- examples/tauri/index.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Welcome to Tauri!</title>
-  </head>
-  <body>
-    <h1>Welcome to Tauri!</h1>
-    <button id="test1">test command</button>
-    <a href="/test.html">go to test</a>
-    <script>
-      window.onload = function () {
-        const rpcApiUrl = window.location.hash.slice(1);
-        const test1 = document.getElementById("test1");
-
-        window.__TAURI__.event.listen("from-window", (event) => {
-          console.log("got from window event", event);
-        });
-
-        test1.onclick = function (e) {
-          e.preventDefault();
-          window
-            .__TAURI__.core.invoke("test1", {
-              url: "https://www.rust-lang.org",
-              timeoutSeconds: Math.floor(Math.random() * 10),
-            })
-            .then(() => {
-              console.log("done");
-            });
-        };
-      };
-    </script>
-  </body>
-</html>
