+++ examples/tauri/src-tauri/.cargo/config.toml
@@ -1,0 +1,4 @@
+[build]
+target = "aarch64-apple-darwin"
+
+[target]
