+++ examples/tauri/src-tauri/.gitignore
@@ -1,0 +1,6 @@
+# Generated by Cargo
+# will have compiled files and executables
+/target/
+
+# mobile projects
+/gen
