+++ examples/tauri/src-tauri/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "tauri-example"
+version = "0.0.0"
+description = "A Tauri App"
+authors = ["you"]
+license = ""
+repository = ""
+edition = "2021"
+publish = false
+
+[lib]
+name = "tauri_example_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
+
+[build-dependencies]
+tauri-build.workspace = true
+
+[dependencies]
+tauri.workspace = true
+tauri-plugin-devtools = { path = "../../../tauri-plugin" }
+tracing.workspace = true
+tokio = { workspace = true, features = ["time"] }
+reqwest = { version = "0.11", default-features = false, features = [
+  "rustls-tls",
+] }
+serde = "1"
+
+[features]
+# this feature is used for production builds or when `devPath` points to the filesystem
+# DO NOT REMOVE!!
+custom-protocol = ["tauri/custom-protocol"]
