+++ examples/tauri/src-tauri/src/main.rs
@@ -1,0 +1,6 @@
+// Prevents additional console window on Windows in release, DO NOT REMOVE!!
+#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
+
+fn main() {
+    tauri_example_lib::run()
+}
