+++ examples/tauri/src/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Welcome to Tauri!</title>
+</head>
+
+<body>
+  <h1>Welcome to Tauri!</h1>
+  <button id="test1">test command</button>
+  <a href="/test.html">go to test</a>
+  <script src="index.js"></script>
+</body>
+
+</html>
