+++ examples/tauri/src/test.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Welcome to Tauri!</title>
+</head>
+
+<body>
+  <h1>Welcome to Tauri!</h1>
+  <a href="/index.html">go back to index</a>
+</body>
+
+</html>
