--- examples/tauri/tauri.conf.json
@@ -1,53 +1,0 @@
-{
-  "$schema": "../../core/tauri-config-schema/schema.json",
-  "build": {
-    "distDir": ["index.html", "test.html"],
-    "devPath": "http://localhost:1420",
-    "beforeDevCommand": "",
-    "beforeBuildCommand": "",
-    "withGlobalTauri": true
-  },
-  "package": {
-    "productName": "Hello World",
-    "version": "0.1.0"
-  },
-  "tauri": {
-    "bundle": {
-      "active": true,
-      "targets": "all",
-      "identifier": "com.tauri.dev",
-      "icon": [
-        ".icons/32x32.png",
-        ".icons/128x128.png",
-        ".icons/128x128@2x.png",
-        ".icons/icon.icns",
-        ".icons/icon.ico"
-      ],
-      "resources": [],
-      "externalBin": [],
-      "copyright": "",
-      "category": "DeveloperTool",
-      "shortDescription": "",
-      "longDescription": "",
-      "deb": {
-        "depends": []
-      },
-      "macOS": {
-        "frameworks": [],
-        "exceptionDomain": ""
-      }
-    },
-    "windows": [
-      {
-        "title": "Welcome to Tauri!",
-        "width": 800,
-        "height": 600,
-        "resizable": true,
-        "fullscreen": false
-      }
-    ],
-    "security": {
-      "csp": "default-src 'self'; connect-src ipc: https://ipc.localhost"
-    }
-  }
-}
