--- examples/tauri/test.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Welcome to Tauri!</title>
-  </head>
-  <body>
-    <h1>Welcome to Tauri!</h1>
-    <a href="/index.html">go back to index</a>
-  </body>
-</html>
